Make useFilterCreatedInField hook
